stmgr: Split stmgr.go into smaller subfiles
